+++ framework/components/pubsub.go
@@ -1,5 +1,7 @@
 package components
 
+import "fmt"
+
 func NewPubsub() *Pubsub {
 	return &Pubsub{
 		ready: make(chan struct{}),
@@ -18,6 +20,10 @@ type Pubsub struct {
 
 func (c *Pubsub) Start() {
 	defer close(c.ready)
+	if containerIsRunning("redis_mtf") {
+		fmt.Printf("pubsub_mtf is already running")
+		return
+	}
 	cmd := `docker run --rm -d --network=mtf_net --name pubsub_mtf --hostname=pubsub_mtf -p 8085:8085 adilsoncarvalho/gcloud-pubsub-emulator`
 	run(cmd)
 }
@@ -27,6 +33,7 @@ func (c *Pubsub) Stop() {
 }
 
 func (c *Pubsub) Ready() {
+	return
 	<-c.ready
 	waitForPortOpen("localhost", "8001")
 }
