+++ framework/components/redis.go
@@ -1,5 +1,7 @@
 package components
 
+import "fmt"
+
 func NewRedis() *Redis {
 	return &Redis{
 		ready: make(chan struct{}),
@@ -16,12 +18,18 @@ type Redis struct {
 }
 
 func (c *Redis) Start() {
+	defer close(c.ready)
+	if containerIsRunning("redis_mtf") {
+		fmt.Printf("mysql_mtf is already running")
+		return
+	}
+
 	cmd := `docker run --rm -d --network=mtf_net --name redis_mtf --hostname=redis_mtf --env REDIS_PASSWORD=test -p 6379:6379 bitnami/redis:4.0`
 	run(cmd)
-	close(c.ready)
 }
 
 func (c *Redis) Stop() {
+	return
 	run("docker stop redis_mtf")
 }
